+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Changes in version 1.18**
 
+- Changed Survival Time algorithm so it returns null values for no-value year/month/day instead of spaces.
 - Updated the SEER Historic Stage algorithm to reflect the algorithm used for the current SEER research data.
 
 **Changes in version 1.17**